Move more state to registry + clean up
