+++ plugin/definition.py
@@ -1,7 +1,6 @@
 import sublime
 
-from .core.clients import LspTextCommand
-from .core.registry import client_for_view
+from .core.registry import client_for_view, LspTextCommand
 from .core.protocol import Request, Point
 from .core.documents import get_document_position, get_position, is_at_word
 from .core.url import uri_to_filename
