+++ plugin/rename.py
@@ -1,6 +1,5 @@
 import sublime_plugin
-from .core.clients import LspTextCommand
-from .core.registry import client_for_view
+from .core.registry import client_for_view, LspTextCommand
 from .core.protocol import Request
 from .core.documents import get_document_position, get_position, is_at_word
